m2/requirements.txt
websocket-client
requests
py-radius
